update links after renaming to skewer-case
